Refactoring: group most Airtable code under lib/airtable
