+++ components/onboarding/form/index.tsx
@@ -2,7 +2,7 @@ import React, { useReducer } from "react";
 import Input from "components/onboarding/form/input";
 import { Button } from "components/buttons";
 import Callout from "components/onboarding/callout";
-import { Field } from "lib/skills";
+import { Field } from "lib/airtable/skills";
 import Strings from "../strings.json";
 import * as Components from "components/onboarding/styles";
 import * as S from "./styles";
