+++ pages/api/protected/me.ts
@@ -1,13 +1,13 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { getToken } from "next-auth/jwt";
-import { send } from "lib/airtable-request";
+import { send } from "lib/airtable/request";
 import { assert } from "console";
 import Airtable from "airtable";
 import {
   getUserProfile,
   updateUserProfile,
   userProfileTable,
-} from "lib/user-profile";
+} from "lib/airtable/user-profile";
 
 /**
  * Retrieve or update user profile
