+++ pages/profile.tsx
@@ -2,8 +2,8 @@ import { NextPage, GetStaticProps } from "next";
 import { useSession, signIn, signOut } from "next-auth/react";
 import { useEffect, useState } from "react";
 import { UserProfilePageState, UserProfilePage } from "components/user-profile";
-import { decodeUserProfile, UserProfile } from "lib/user-profile";
-import { flattenSkills, Skill } from "lib/skills";
+import { decodeUserProfile, UserProfile } from "lib/airtable/user-profile";
+import { flattenSkills, Skill } from "lib/airtable/skills";
 import { siteData } from "lib/site-data";
 
 type PageProps = {
